// src/app/interfaces/PharmRequest.ts
interface PharmRequest {
    name: string;
    nameOwner: string;
    address: string;
    email: string;
    password: string;
    RUC: number;
    phone: number;
    photo: string;
    razonSocial: string;
    latitude: number;
    length: number;
    _id: string;
}

export default PharmRequest;